+++ orm/model/ormdb/module_test.go
@@ -16,11 +16,10 @@ import (
 	appv1alpha1 "cosmossdk.io/api/cosmos/app/v1alpha1"
 	ormmodulev1alpha1 "cosmossdk.io/api/cosmos/orm/module/v1alpha1"
 	ormv1alpha1 "cosmossdk.io/api/cosmos/orm/v1alpha1"
-	"cosmossdk.io/core/appconfig"
-	"cosmossdk.io/core/appmodule"
 	"cosmossdk.io/core/genesis"
 	"cosmossdk.io/core/store"
 	"cosmossdk.io/depinject"
+	"cosmossdk.io/depinject/appconfig"
 	_ "cosmossdk.io/orm" // required for ORM module registration
 	"cosmossdk.io/orm/internal/testkv"
 	"cosmossdk.io/orm/internal/testpb"
@@ -36,8 +35,8 @@ import (
 
 func init() {
 	// this registers the test module with the module registry
-	appmodule.Register(&testpb.Module{},
-		appmodule.Provide(NewKeeper),
+	appconfig.RegisterModule(&testpb.Module{},
+		appconfig.Provide(NewKeeper),
 	)
 }
 
@@ -362,11 +361,11 @@ type testStoreService struct {
 }
 
 func (t testStoreService) OpenKVStore(context.Context) store.KVStore {
-	return t.db
+	return testkv.TestStore{Db: t.db}
 }
 
 func (t testStoreService) OpenMemoryStore(context.Context) store.KVStore {
-	return t.db
+	return testkv.TestStore{Db: t.db}
 }
 
 func TestGetBackendResolver(t *testing.T) {