+++ x/staking/autocli.go
@@ -174,6 +174,13 @@ func (am AppModule) AutoCLIOptions() *autocliv1.ModuleOptions {
 					PositionalArgs: []*autocliv1.PositionalArgDescriptor{{ProtoField: "validator_address"}, {ProtoField: "amount"}, {ProtoField: "creation_height"}},
 				},
 				{
+					RpcMethod:      "RotateConsPubKey",
+					Use:            "rotate-cons-pubkey [validator-address] [new-pubkey]",
+					Short:          fmt.Sprintf("rotate validator consensus pub key. Note: you have to replace the `~/.%sd/config/priv_validator_key.json` with new key and restart the node after rotating the key", version.AppName),
+					Example:        fmt.Sprintf(`%s tx staking rotate-cons-pubkey myvalidator {"@type":"/cosmos.crypto.ed25519.PubKey","key":"oWg2ISpLF405Jcm2vXV+2v4fnjodh6aafuIdeoW+rUw="}`, version.AppName),
+					PositionalArgs: []*autocliv1.PositionalArgDescriptor{{ProtoField: "validator_address"}, {ProtoField: "new_pubkey"}},
+				},
+				{
 					RpcMethod:      "UpdateParams",
 					Use:            "update-params-proposal [params]",
 					Short:          "Submit a proposal to update staking module params. Note: the entire params must be provided.",